Implemented `infrastructure string`
This is not a Ruby-esque String, it’s much more like a cross between `NSString` and Ruby’s `Symbol` class.

It essentially operates as a globally-unique name. It is intended to be implemented in much-faster ways in other implementations; here, it is fairly inefficient.

A mutable `infrastructure soup` will eventually be provided for more generalized text operations.