Renamed `bool.null` to `bool.void`.
`void` is a clearer meaning; it’s a value into which lookups and calls fall:

null |nəl| 2 having or associated with the value zero.

void |void| 1 a completely empty space : the black void of space; an emptiness caused by the loss of something : the void left by the collapse of communism.