(- fix client closes:#3) runInNewContext wasn't calling into the DOM API properly.
`Node#insertBefore()` requires a second argument, that being *which element* to insert the element
before in the parent's NodeList. Most browsers were lenient and assumed the second argument to be
`null` if none is passed (which *also* meant we were unintentionally appending the element to the
*end* of the node-list, which is exactly the opposite of what we wanted.)

This mostly affected FireFox, it seems; everybody else was more lenient on requring the second
argument.