(meta fix api) Changing data.coffee from API-modification to exports
For reasons that I can't discern, the coverage system was choking on the old format, here. For
whatever reason, when executed via CoffeeScript, everything worked fine; but when compiled to JS,
`data.coffee` wasn't being loaded in time for `reactor.coffee` to have the data-types made available
to it at require-time. Very strange.

This is probably slightly more idiomatic *anyway*, so a net benefit.