(re new API) Implement Operation type (and registry), for queue
The new `Operation` type forms the initial element of the new queueless reactor. For now, I've only
implemented the old `advance` method as an `Operation`; but soon ownership will also be
re-implemented, which will involve an `adopt` op. (The actual implementation of the operation-queue
for `Execution`s is forthcoming.)

In addition to all the obvious changes (loss of `reactor.*`) and re-implementations, I should note
that `Execution#register_response` has become `Execution#respond`.

/cont:fb99ef45 squashes:"5db18d01..f13a8b0e, ff0608dc"