(!! new API inc!) Implement responsibility!
This includes a whole slew of changes.

I entirely ripped out `LiabilityFamily`, for one. It's going to be so difficult to maintain LiabilityFamilies (imagine: owned-subgraph splits in the middle, with some Liabilities rooted at the top of the subgraph, but some Liabilities rooted below the split …), and I don't see any purpose in them besides centralizing the write/read state-and-logic.

This also includes the first attempt at `::walk`, which would eventually become Giraphe.

/squashes "65d9969b81..fa17ee7933"